Refactor files and modify compose.yml
