+++ package.json
@@ -4,9 +4,7 @@
   "private": true,
   "scripts": {
     "build": "next build",
-    "dev:amqpserver": "ts-node src/server/matchingServer.ts",
-    "dev:client": "next dev",
-    "dev": "concurrently \"yarn run dev:amqpserver\" \"yarn run dev:client\"",
+    "dev": "next dev",
     "postinstall": "yarn prisma:generate",
     "lint": "next lint",
     "start": "next start",
@@ -40,8 +38,6 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "sass": "^1.67.0",
-    "tls": "^0.0.1",
-    "ws": "^8.14.1",
     "rehype-sanitize": "^6.0.0",
     "rehype-stringify": "^10.0.0",
     "remark-gfm": "^3.0.1",
@@ -58,7 +54,6 @@
     "@types/node": "^18.16.0",
     "@types/react": "^18.2.20",
     "@types/react-dom": "^18.2.7",
-    "@types/ws": "^8.5.5",
     "@types/s3rver": "^3.7.0",
     "@typescript-eslint/eslint-plugin": "^6.3.0",
     "@typescript-eslint/parser": "^6.3.0",
@@ -70,7 +65,7 @@
     "prettier-plugin-tailwindcss": "^0.5.1",
     "prisma": "^5.1.1",
     "s3rver": "^3.7.1",
-    "sass": "^1.66.1",
+    "sass": "^1.67.0",
     "tailwindcss": "^3.3.3",
     "typescript": "^5.1.6"
   },
