+++ src/pages/_app.tsx
@@ -4,7 +4,6 @@ import { type AppType } from "next/app";
 
 import { api } from "~/utils/api";
 
-import "~/styles/globals.css";
 import "~/styles/globals.sass";
 
 const MyApp: AppType<{ session: Session | null }> = ({
